% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\alias{multiply}
\title{basic arithmetic}
\usage{
add(x, y)

multiply(x, y)
}
\arguments{
\item{x, y}{numbers}
}
\value{
\code{add} sum of and y \code{multiply} multiple of x and y
}
\description{
Return sum of two numbers
}
\examples{
add(1,2)
multiply(1,2)
}
