% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{as.Classif.table}
\alias{as.Classif.table}
\title{Create a Classification table.}
\usage{
as.Classif.table(y_pred, y_truth, ..., p.threshold = 0.5)
}
\arguments{
\item{y_pred}{Predicted probabilities for the positive class.}

\item{y_truth}{Vector containing the ground truth of positive class.}

\item{p.threshold}{(optional) Threshold to determine label.}
}
\value{
list containing a confusion matrix and quick accessors.
}
\description{
Create a Classification table.
}
