% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{as.Confusion.mat}
\alias{as.Confusion.mat}
\title{Create a confusion matrix.}
\usage{
as.Confusion.mat(y_pred, y_truth, ..., p.threshold = 0.5)
}
\arguments{
\item{y_pred}{Predicted probabilities for the positive class.}

\item{y_truth}{Vector containing the ground truth of positive class.}

\item{p.threshold}{(optional) Threshold to determine label.}
}
\value{
2x2 confusion matrix.
}
\description{
Create a confusion matrix.
}
