% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.ACC}
\alias{calc.ACC}
\title{Calculate the accuracy of a model.}
\usage{
calc.ACC(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(TP + TN) / (TP + FP + FN + TN)
}
\description{
Calculate the accuracy of a model.
}
