% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.CohenKappaScore}
\alias{calc.CohenKappaScore}
\title{Calculate the Cohen Kappa metric}
\usage{
calc.CohenKappaScore(fit.mat, random.mat)
}
\arguments{
\item{fit.mat}{2x2 confusion matrix for fit model.}

\item{random.mat}{2x2 confusion matrix for random classifier.}
}
\value{
k = (ACC(fit) - ACC(random)) / (1 - ACC(random))
}
\description{
Calculate the Cohen Kappa metric
}
