% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.ERR}
\alias{calc.ERR}
\title{Calculate the error rate of the model.}
\usage{
calc.ERR(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix.}
}
\value{
(FP + FN) / (TP + TN + FN + FP)
}
\description{
Calculate the error rate of the model.
}
