% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.F2}
\alias{calc.F2}
\title{Calculate the F2 score.}
\usage{
calc.F2(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
((2 x Recall x Precision) / (Recall + Precision))
}
\description{
Calculate the F2 score.
}
