% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.FDR}
\alias{calc.FDR}
\title{Calculate the false discovery rate.}
\usage{
calc.FDR(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(FP) / (TP + FP)
}
\description{
Calculate the false discovery rate.
}
