% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.FNR}
\alias{calc.FNR}
\title{Calculate the false negative rate.}
\usage{
calc.FNR(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(FN) / (TP + FN)
}
\description{
Calculate the false negative rate.
}
