% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.FPR}
\alias{calc.FPR}
\title{Calculate the false positive rate.}
\usage{
calc.FPR(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(FP) / (FP + TN)
}
\description{
Calculate the false positive rate.
}
