% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.FScore}
\alias{calc.FScore}
\title{Calculate the F-Score of a model.}
\usage{
calc.FScore(mat, ..., beta = 1)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(1 + beta^2) x ((precision x recall) / (beta^2 x precision) + recall)
}
\description{
Calculate the F-Score of a model.
}
