% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.NEG}
\alias{calc.NEG}
\title{Given a confusion matrix, count the negative classifications.}
\usage{
calc.NEG(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix.}
}
\value{
FP + TN
}
\description{
Given a confusion matrix, count the negative classifications.
}
