% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.NPV}
\alias{calc.NPV}
\title{Calculate the negative predictive value.}
\usage{
calc.NPV(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(TN) / (TN + FN)
}
\description{
Calculate the negative predictive value.
}
