% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.POS}
\alias{calc.POS}
\title{Given a confusion matrix, count the positive classifications.}
\usage{
calc.POS(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix.}
}
\value{
TP + FN
}
\description{
Given a confusion matrix, count the positive classifications.
}
