% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.PPV}
\alias{calc.PPV}
\title{Calculate the positive predictive value}
\usage{
calc.PPV(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(TP) / (TP + FP)
}
\description{
Calculate the positive predictive value
}
