% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.Precision}
\alias{calc.Precision}
\title{Calculate the precision of a model.}
\usage{
calc.Precision(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(TP) / (TP + FP)
}
\description{
Calculate the precision of a model.
}
\seealso{
calc.PPV
}
