% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.Recall}
\alias{calc.Recall}
\title{Calculate the recall of a model.}
\usage{
calc.Recall(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(TP) / (TP + FN)
}
\description{
Calculate the recall of a model.
}
