% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.TNR}
\alias{calc.TNR}
\title{Calculate the true negative rate.}
\usage{
calc.TNR(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(TN) / (TN + FP)
}
\description{
Calculate the true negative rate.
}
