% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.TOTAL}
\alias{calc.TOTAL}
\title{Given a confusion matrix, count the total classifications.}
\usage{
calc.TOTAL(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix.}
}
\value{
TP + TN + FN + FP = P + N = TOTAL
}
\description{
Given a confusion matrix, count the total classifications.
}
