% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{calc.TPR}
\alias{calc.TPR}
\title{Calculate the true positive rate.}
\usage{
calc.TPR(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix}
}
\value{
(TP) / (TP + FN)
}
\description{
Calculate the true positive rate.
}
