% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{elnet_coord}
\alias{elnet_coord}
\title{This function fits elastic net to data using coordinate descent algorithm.}
\usage{
elnet_coord(
  X,
  Y,
  lambda = 0.1,
  alpha = 0.5,
  max_it = 10^6,
  tolerance = 10^(-4)
)
}
\arguments{
\item{X}{The design matrix}

\item{Y}{The response vector}

\item{lambda}{the hyperparameter for the restriction term(s).
By default it's 0.5.}

\item{alpha}{#the hyperparameter of the elastic net. By default alpha = 0,
which is of the form of ridge regression.}

\item{max_it}{# the maximum number of iteration. By default it's 10^4}

\item{tolerance}{# return when difference between two step of beta are
smaller then this value.}
}
\value{
# the estimates of the betas
}
\description{
This function fits elastic net to data using coordinate descent algorithm.
}
\examples{
p = 20
n  = 50
beta_0 = rep(0, p)
alpha = c(0, 0.5, 1)
beta_partial = c(2, 0,-2, 0, 1, 0, -1, 0)
beta = c(beta_partial, rep(0, 12))
corr_mat = diag(20)
corr_mat[1, 2] = 0.8
corr_mat[2, 1] = 0.8
corr_mat[5, 6] = 0.8
corr_mat[6, 5] = 0.8
eps = matrix(rnorm(n), n)
X = MASS::mvrnorm(n, rep(0, p), Sigma = corr_mat)
Y = X[,] \%*\% beta + eps
elnet_coord(X,Y)
print(elnet_coord(X,Y))
}
