% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{example_se}
\alias{example_se}
\title{Example fission yeast SummarizedExperiment}
\format{
A SummarizedExperiment object with:
\describe{
\item{assays}{counts - raw count matrix}
\item{colData}{sample metadata including strain and minute}
}
}
\source{
fission Bioconductor package
}
\usage{
example_se
}
\description{
RNA-seq count data from the fission yeast stress response experiment.
Includes sample metadata such as strain and minute.
}
\keyword{datasets}
