% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator.R
\name{f_operator}
\alias{f_operator}
\title{transform \emph{sqrt.2()} and \emph{log.2()} into functions that can handle errors}
\usage{
f_operator(f)
}
\arguments{
\item{f}{function log.2 or sqrt.2}
}
\value{
a function that can handle error return error object
}
\description{
transform \emph{sqrt.2()} and \emph{log.2()} into functions that can handle errors
}
\examples{
sqrt.3 = f_operator(sqrt.2)
set.seed(2)
(y = lapply(runif(2,-10,10), sqrt.3))
}
