% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator_log.R
\name{f_operator_log}
\alias{f_operator_log}
\title{Allow applying error-giving log}
\usage{
f_operator_log(f)
}
\arguments{
\item{f}{a function that takes a numeric value}
}
\value{
transformed f to catch_cnd
}
\description{
Allow applying error-giving log
}
\examples{
g <- f_operator_log(exp)
g(5)
}
