% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator_sqrt.R
\name{f_operator_sqrt}
\alias{f_operator_sqrt}
\title{Allow applying error giving sqrt}
\usage{
f_operator_sqrt(f)
}
\arguments{
\item{f}{a function that takes numeric value input}
}
\value{
transformed f to catch_cnd
}
\description{
Allow applying error giving sqrt
}
\examples{
g <- f_operator_sqrt(exp)
g(5)
}
