% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{FN}
\alias{FN}
\title{Given a confusion matrix, get the false negative (type II error).}
\usage{
FN(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix.}
}
\value{
value representing false negative
}
\description{
Given a confusion matrix, get the false negative (type II error).
}
