% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{FP}
\alias{FP}
\title{Given a confusion matrix, get the false positive (type I error).}
\usage{
FP(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix.}
}
\value{
value representing false positive.
}
\description{
Given a confusion matrix, get the false positive (type I error).
}
