% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_parallel.R
\name{gen_A}
\alias{gen_A}
\title{Generate a tridiagonal matrix}
\usage{
gen_A(n, alpha)
}
\arguments{
\item{n}{the size of the matrix}

\item{alpha}{the value along the diagonal}
}
\value{
Returns A, an n x n tridiagonal matrix
}
\description{
Generates a tridiagonal matrix with ones on the offdiagonals and alpha on the diagonal
}
