% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{gen_dat}
\alias{gen_dat}
\title{A helper function to generate data for algorithmic leveraging}
\usage{
gen_dat(n, df = 6)
}
\arguments{
\item{n}{number of data points}

\item{df}{degrees of freedom for t-distribution over x. Default=6}
}
\value{
Returns list(x, y)
}
\description{
Generates noisey x,y data where y = -x + noise and x ~ t-distribution
}
\examples{
gen_dat(n=10, df=6)
}
