% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{gen_elnet_data}
\alias{gen_elnet_data}
\title{Generates Data like used in HW1 to run elnet_coord}
\usage{
gen_elnet_data(n)
}
\arguments{
\item{n}{number of data points to use}
}
\value{
Returns list(x, y)
}
\description{
generates X (n x 20) and y (n x 1) with fixed true betas
}
\examples{
dat = gen_elnet_data(20)
X = dat[[1]]
y = dat[[2]]
}
