% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{install_ads8192_cli}
\alias{install_ads8192_cli}
\title{Install ADS8192 CLI launchers}
\usage{
install_ads8192_cli()
}
\value{
Invisible NULL
}
\description{
Installs the package CLI launcher(s) onto the user's PATH using Rapp.
}
