% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_parallel.R
\name{jacobi_parallel}
\alias{jacobi_parallel}
\title{Parallel-Processed Jacobi Method}
\usage{
jacobi_parallel(A, b, n_iters = 100, ncores = 1)
}
\arguments{
\item{A}{Input Matrix for Ax=b}

\item{b}{right-hand side vector for Ax=b}

\item{n_iters}{the number of jacobi iterations to run. Degault=100}

\item{ncores}{the number of cores to use. Default=1}
}
\value{
Returns x, the approximate solution to Ax=b
}
\description{
Numerically approximates solutions to linear systems like Ax=b
}
\examples{
A = gen_A(100, 3)
x = as.matrix(rep(c(1,0), 50))
b = A \%*\% v
jacobi_parallel(A,b, n_iters=100, n_cores=1) # returns approximation to x
}
