% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver_data}
\alias{liver_data}
\title{Liver malfunction dataset}
\format{
A data frame with seven variables:
\describe{
\item{\code{blood.1}}{Quantitative results from undisclosed blood test related
to liver function.}
\item{\code{blood.2}}{Quantitative results from undisclosed blood test related
to liver function.}
\item{\code{blood.3}}{Quantitative results from undisclosed blood test related
to liver function.}
\item{\code{blood.4}}{Quantitative results from undisclosed blood test related
to liver function.}
\item{\code{blood.5}}{Quantitative results from undisclosed blood test related
to liver function.}
\item{\code{drinks}}{Number of alcoholic beverages reported as consumed.}
\item{\code{severity}}{Severity of liver malfunction. Group 1 indicates severe
malfunction.}
}
}
\source{
Data provided by course instructor.
}
\usage{
liver_data
}
\description{
Observations on 345 men in detailed report on liver malfunctions.
}
\keyword{datasets}
