% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log.2}
\alias{log.2}
\title{Calculate log of a function}
\usage{
\method{log}{`2`}(x)
}
\arguments{
\item{x}{the input value}
}
\value{
the log value of x
}
\description{
Calculate log of a function
}
\examples{
log(2)
}
