% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{make.Confusion.mat}
\alias{make.Confusion.mat}
\title{Create a confusion matrix explicitly.}
\usage{
make.Confusion.mat(..., tp, fp, fn, tn)
}
\arguments{
\item{tp}{true positive}

\item{fp}{false positive (Type I error)}

\item{fn}{false negative (Type II error)}

\item{tn}{true negative}
}
\value{
2x2 confusion matrix.
}
\description{
Create a confusion matrix explicitly.
}
