% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{map.where}
\alias{map.where}
\title{Map probability vector to positive and negative values if row meets
or exceeds a particular threshold.}
\usage{
map.where(condition, ..., pos = TRUE, neg = FALSE)
}
\arguments{
\item{pos}{Positive class label.}

\item{neg}{Negative class label.
#'}

\item{y_pred}{Probabilities to map.}
}
\value{
vector of mapped values.
}
\description{
Map probability vector to positive and negative values if row meets
or exceeds a particular threshold.
}
