% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{my_glmnet}
\alias{my_glmnet}
\title{a variant of glmnet}
\usage{
my_glmnet(covariate, response, ...)
}
\arguments{
\item{covariate}{the model matrix}

\item{response}{the response vector}

\item{...}{other parameter passed to \code{glmnet}}
}
\value{
a glmnet model
}
\description{
a variant of glmnet
}
\examples{
#no examples
}
\seealso{
\link[glmnet]{glmnet}
}
