% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{severity_map}
\alias{severity_map}
\title{Severity group mapping}
\format{
A data frame:
\describe{
\item{\code{status}}{Human-readable name for the severity group.}
\item{\code{value}}{Group number used in original dataset.}
\item{\code{code}}{Mapped class label for severity status..}
}
}
\source{
Table generated by \code{data-raw/severity_map.R}
}
\usage{
severity_map
}
\description{
Lookup table containing original severity groups mapped to the
positive and negative class labels.
}
\keyword{datasets}
