% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt2.R
\name{sqrt.2}
\alias{sqrt.2}
\title{Calculate sqrt
This is a test function in HW2
Molin Yue}
\usage{
\method{sqrt}{`2`}(x)
}
\arguments{
\item{x}{The input number}
}
\value{
the square root of your input number
}
\description{
Calculate sqrt
This is a test function in HW2
Molin Yue
}
\examples{
sqrt.2(1)
sqrt.2(2)
}
\seealso{
log2
}
