% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt2.R
\name{sqrt2}
\alias{sqrt2}
\title{sqrt that gives error not warning}
\usage{
sqrt2(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
sqrt of x
}
\description{
sqrt that gives error not warning
}
\examples{
#
sqrt2(25)
}
