% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{TN}
\alias{TN}
\title{Given a confusion matrix, get the true negative.}
\usage{
TN(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix.}
}
\value{
value representing true negative.
}
\description{
Given a confusion matrix, get the true negative.
}
