% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confmat.R
\name{TP}
\alias{TP}
\title{Given a confusion matrix, get the true positive.}
\usage{
TP(mat)
}
\arguments{
\item{mat}{2x2 confusion matrix.}
}
\value{
value representing true positive.
}
\description{
Given a confusion matrix, get the true positive.
}
