% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Algorithmic Leveraging}
\usage{
algo_leverage(X, Y, algorithm = "leverage", sub_rows = 100)
}
\arguments{
\item{X}{The design matrix}

\item{Y}{The response variable vector}

\item{algorithm}{Including "uniform" and "leverage", user can choose to
use 1 and 2, respectively，for these two choices. Default setting is "leverage".}

\item{sub_rows}{Number of subsamples to generate. Default setting is
100.}
}
\value{
The parameter estimators.
}
\description{
This function implements algorithmic leveraging for
linear regression using uniform and leverage score, and the user can
specific subsampling of rows.
}
\examples{
n = 500
X = matrix(rt(n, 6),n)
eps = rnorm(n)#error term
Y = - X + eps
algo_leverage(X,Y)
}
