% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Algorithmic Leveraging for Linear Regression}
\usage{
algo_leverage(x, y, subsample = "unif", r = 0)
}
\arguments{
\item{x}{predictor variable}

\item{y}{response variable}

\item{subsample}{Should subsamples be taken uniformly or according to leverage
scores? Options are 'unif' and 'lev'. Default='unif'}

\item{r}{Size of subsample, defaults to ~10% of data}
}
\value{
Returns lm class predicted
}
\description{
Computes estimate of SLR coefficient using subsampling
}
\examples{
algo_leverage(x, y, subsample='lev', r=10)
}
