% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Solving OLS using subsampling}
\usage{
algo_leverage(X, y, r, type)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{r}{number of subsamples}

\item{type}{uniform weightage or leverage weightage (default uniform)}
}
\value{
coefficient of the regression
}
\description{
Solving OLS using subsampling
}
