% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{algo_leverage}
\usage{
algo_leverage(A, b, n_samples)
}
\arguments{
\item{A}{Design matrix}

\item{b}{Target vector}

\item{n_samples}{Number of samples to use for leverage sampling.}
}
\description{
Use leverage sampling to solve least square problem.  Returns the intercept and slope
}
\examples{
A = rt(100,6)
e = rnorm(100, 0,1)
b = -A+e
res = algo_leverage(A,b,100)
print(res)
}
