% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{algo_leverage Function}
\usage{
algo_leverage()
}
\description{
This function implements algorithmic leveraging for linear regression using uniform and leverage score based subsampling of rows.
More specifically, the simulation setting of Figure 1 in https://onlinelibrary.wiley.com/doi/full/10.1002/wics.1324 is considered.
}
