% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{algo_leverage Function}
\usage{
algo_leverage()
}
\value{
a list contains two numeric vectors, the simulated design matrix \eqn{X} and the simulated response matrix \eqn{Y}. 
The two numeric vectors corresponding to uniform sampling (beta_unif) and leverage sampling (beta_lev). 
Each vector contains regression coefficients obtained using r = 10, 50, 100, 200, 300 in order.
}
\description{
This function implements algorithmic leveraging for linear regression using uniform and leverage score based subsampling of rows.
More specifically, the simulation setting of Figure 1 in https://onlinelibrary.wiley.com/doi/full/10.1002/wics.1324 is considered.
The function will automatically generate the simulation setting in previous paper, so it does not take any arguments.
}
