% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{Algorithmic Leveraging for Linear Regression}
\usage{
elnet_coord(X, y, alpha = 0.5, lambda = 1, n_iters = 100)
}
\arguments{
\item{y}{response variable (n x 1)}

\item{alpha}{Ratio of elastic net (alpha=1 corresponds to lasso)}

\item{lambda}{regularization weight. Default=1}

\item{n_iters}{number of update iterations. Default=100}

\item{x}{predictor matrix (n x p)}
}
\value{
Returns regression estimates (p x 1)
}
\description{
Computes estimate of SLR coefficient using subsampling
}
\examples{
dat = elnet_data_gen(n)
X = dat[[1]]
y = dat[[2]]
elnet_coord(x, y, alpha=0.8, lambda=2)
}
