% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.r
\name{elnet_coord}
\alias{elnet_coord}
\title{Solving elastic net via coordinate descent}
\usage{
elnet_coord(y, X, Lambda, alpha, beta_init, iterlength)
}
\arguments{
\item{y}{The response vector}

\item{X}{Design matrix}

\item{Lambda}{Sequence of penalizing weights}

\item{alpha}{weightage on L_1 norm}

\item{beta_init}{initial coefficient vector (default rep(0,ncol(X)) - the zero vector )}

\item{iterlength}{no. of iteration of the coordinate descent algorithm (default = 1000)}
}
\value{
Solution path - A matrix of dimension (length(Lambda),ncol(X)) where each row gives the coefficient vector for a fixed Lambda.
}
\description{
Solving elastic net via coordinate descent
}
