% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.r
\name{elnet_coord}
\alias{elnet_coord}
\title{Solving elastic net via coordinate descent}
\arguments{
\item{y}{The response vector}

\item{X}{Design matrix}

\item{Lambda}{Sequence of penalizing weights}

\item{alpha}{weightage on L_1 norm}

\item{beta_init}{initial coefficient vector}

\item{iterlength}{no. of iteration of the coordinate descent algorithm}
}
\value{
None
}
\description{
Solving elastic net via coordinate descent
}
