% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{elnet_coord Function}
\usage{
elnet_coord()
}
\description{
This function  fits elastic net to data using coordinate descent algorithm. The simulation setting is specified in homework 1 problem 3
part 2.
}
