% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{Plot mCPR data and estimates}
\usage{
plot_cp(dat, est, iso_code, CI = 95)
}
\arguments{
\item{dat:}{tibble which contains mCPR observations. Columns: iso, cpr_year, mCPR}

\item{est:}{tibble which contains mCPR estimates. Columns: “Country or area”, iso, Year, Median, U95, L95}

\item{iso_code:}{country iso code}

\item{CI:}{Confidence intervals to be plotted. Options can be: 80, 95, or NA (no CI plotted)}
}
\value{
: ggplot object with data and estimates
}
\description{
Plot mCPR data and estimates
}
\examples{
function(dat, est, iso_code, CI=95)
}
