% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{ploting a Country
This function is the same as in HW1 just in a package}
\usage{
plot_cp(dat, est, iso_code, CI = 95)
}
\arguments{
\item{dat}{the observed data for modern contraceptive use}

\item{est}{the estimated data for modern contraceptive use}

\item{iso_code}{The ISO Code of the country you are searching for}

\item{CI}{default confidence intervall is 95\% it can also take an 80\% and NA}
}
\value{
The ggplot that will be created by the function
}
\description{
ploting a Country
This function is the same as in HW1 just in a package
}
\examples{
plot_cp(dat_example, est_example, iso_code = 4, CI = 95)
}
