% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{Solving linear systems of the form Ax=b}
\usage{
solve_ols(A, b, method, tol, maxiter, core)
}
\arguments{
\item{A}{matrix of the system to be solved}

\item{b}{vector of the system to be solved}

\item{method}{Gauss, Jacobi or Jacobi parallel}

\item{tol}{tolerance level of relative errors (default 1e-10)}

\item{maxiter}{maximum number of iterations (default 10000)}

\item{core}{number of core used for parallel computing (only valid for Jacobi Parallel)}
}
\value{
The solution of the system
}
\description{
Solving linear systems of the form Ax=b
}
