% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{Solve OLS}
\usage{
solve_ols(A, b, method = "Jacobi", n_cores = 1, iter = 10)
}
\arguments{
\item{A}{design matrix}

\item{b}{target vector}

\item{method}{Jacobi(Jacobi), or Gauss-Seidal(GS)}

\item{n_cores}{Number of parallel processes}

\item{iter}{number of iteration for the iterative method}
}
\value{
list(x, run_time)
}
\description{
Solve ordinary least square using iterative methods
}
\examples{
A = diag(2, 10, 10)
b = 1:10
Jacobi_res = solve_ols(A,b,method="Jacobi",n_cores=2)
GS_res = solve_ols(A,b,method="GS")
print(Jacobi_res[1])
print(GS_res[1])
}
