% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{solve_ols Function}
\usage{
solve_ols(X, b, method, tol = 1e-04)
}
\arguments{
\item{X}{the coefficient matrix of the linear system and should be a square matrix.}

\item{b}{the column vector collects values of each linear equation.}

\item{method}{the method is used to solve the linear system, either "Gauss" or "Jacobi" coresponding to Gauss-Seidel or Jacobi method respectively.}

\item{tol}{the convergence criterion which is measured by the \eqn{l_2} norm of the difference of between the current solution and the solution in the last iteration.}
}
\value{
The solution vector of the linear system specified in the function
}
\description{
This function solves a linear system using Gauss-Seidel or Jacobi method.
}
